+++ geonode/security/models.py
@@ -18,6 +18,7 @@
 #
 #########################################################################
 
+from django.conf import settings
 from django.contrib.auth.models import User, Permission
 from django.contrib.contenttypes.models import ContentType
 from django.contrib.contenttypes.generic import GenericForeignKey
@@ -25,7 +26,8 @@ from django.db import models
 from django.utils.translation import ugettext_lazy as _
 from django.core.exceptions import ObjectDoesNotExist
 from django.contrib.auth import login
-from geonode.contrib.groups.models import Group
+if "geonode.contrib.groups" in settings.INSTALLED_APPS:
+    from geonode.contrib.groups.models import Group
 from geonode.security.enumerations import GENERIC_GROUP_NAMES
 from geonode.security.enumerations import AUTHENTICATED_USERS, ANONYMOUS_USERS
  
@@ -115,22 +117,25 @@ class GroupObjectRoleMapping(models.Model):
     in the context of a specific object.
     """
 
-    group = models.ForeignKey(Group, related_name="role_mappings")
+    if "geonode.contrib.groups" in settings.INSTALLED_APPS:
+        group = models.ForeignKey(Group, related_name="role_mappings")
     
-    object_ct = models.ForeignKey(ContentType)
-    object_id = models.PositiveIntegerField()
-    object = GenericForeignKey('object_ct', 'object_id')
+        object_ct = models.ForeignKey(ContentType)
+        object_id = models.PositiveIntegerField()
+        object = GenericForeignKey('object_ct', 'object_id')
 
-    role = models.ForeignKey(ObjectRole, related_name="group_mappings")
+        role = models.ForeignKey(ObjectRole, related_name="group_mappings")
 
-    def __unicode__(self):
-        return u"%s | %s -> %s" % (
-            unicode(self.object),
-            unicode(self.group), 
-            unicode(self.role))
+        def __unicode__(self):
+            return u"%s | %s -> %s" % (
+                unicode(self.object),
+                unicode(self.group), 
+                unicode(self.role))
 
-    class Meta:
-        unique_together = (('group', 'object_ct', 'object_id', 'role'), )
+        class Meta:
+            unique_together = (('group', 'object_ct', 'object_id', 'role'), )
+    else:
+        pass
 
 
 class PermissionLevelError(Exception):
@@ -190,37 +195,38 @@ class PermissionLevelMixin(object):
             # grant new level
             UserObjectRoleMapping.objects.create(user=user, object=self, role=role)
 
-    def get_group_level(self, group):
-        """
-        get the permission level (if any) specifically assigned to the given group.
-        Returns LEVEL_NONE to indicate no specific level has been assigned.
-        """
-        my_ct = ContentType.objects.get_for_model(self)
-        try:
-            mapping = GroupObjectRoleMapping.objects.get(group=group, object_id=self.id, object_ct=my_ct)
-            return mapping.role.codename
-        except GroupObjectRoleMapping.DoesNotExist:
-            return self.LEVEL_NONE
-
-    def set_group_level(self, group, level):
-        """
-        set the group's permission level to the level specified. if 
-        level is LEVEL_NONE, any existing level assignment is removed.
-        """
-        
-        my_ct = ContentType.objects.get_for_model(self)
-        if level == self.LEVEL_NONE:
-            GroupObjectRoleMapping.objects.filter(group=group, object_id=self.id, object_ct=my_ct).delete()
-        else:
-            # lookup new role...
+    if "geonode.contrib.groups" in settings.INSTALLED_APPS:
+        def get_group_level(self, group):
+            """
+            get the permission level (if any) specifically assigned to the given group.
+            Returns LEVEL_NONE to indicate no specific level has been assigned.
+            """
+            my_ct = ContentType.objects.get_for_model(self)
             try:
-                role = ObjectRole.objects.get(codename=level, content_type=my_ct)
-            except ObjectRole.NotFound: 
-                raise PermissionLevelError("Invalid Permission Level (%s)" % level)
-            # remove any existing mapping              
-            GroupObjectRoleMapping.objects.filter(group=group, object_id=self.id, object_ct=my_ct).delete()
-            # grant new level
-            GroupObjectRoleMapping.objects.create(group=group, object=self, role=role)
+                mapping = GroupObjectRoleMapping.objects.get(group=group, object_id=self.id, object_ct=my_ct)
+                return mapping.role.codename
+            except GroupObjectRoleMapping.DoesNotExist:
+                return self.LEVEL_NONE
+
+        def set_group_level(self, group, level):
+            """
+            set the group's permission level to the level specified. if 
+            level is LEVEL_NONE, any existing level assignment is removed.
+            """
+        
+            my_ct = ContentType.objects.get_for_model(self)
+            if level == self.LEVEL_NONE:
+                GroupObjectRoleMapping.objects.filter(group=group, object_id=self.id, object_ct=my_ct).delete()
+            else:
+                # lookup new role...
+                try:
+                    role = ObjectRole.objects.get(codename=level, content_type=my_ct)
+                except ObjectRole.NotFound: 
+                    raise PermissionLevelError("Invalid Permission Level (%s)" % level)
+                # remove any existing mapping              
+                GroupObjectRoleMapping.objects.filter(group=group, object_id=self.id, object_ct=my_ct).delete()
+                # grant new level
+                GroupObjectRoleMapping.objects.create(group=group, object=self, role=role)
 
     def get_gen_level(self, gen_role):
         """
@@ -259,9 +265,10 @@ class PermissionLevelMixin(object):
         ct = ContentType.objects.get_for_model(self)
         return UserObjectRoleMapping.objects.filter(object_id = self.id, object_ct = ct)
 
-    def get_group_levels(self):
-        ct = ContentType.objects.get_for_model(self)
-        return GroupObjectRoleMapping.objects.filter(object_id = self.id, object_ct = ct)
+    if "geonode.contrib.groups" in settings.INSTALLED_APPS:
+        def get_group_levels(self):
+            ct = ContentType.objects.get_for_model(self)
+            return GroupObjectRoleMapping.objects.filter(object_id = self.id, object_ct = ct)
 
     def get_generic_levels(self):
         ct = ContentType.objects.get_for_model(self)
@@ -298,11 +305,12 @@ class PermissionLevelMixin(object):
             levels[rm.subject] = rm.role.codename
         levels['users'] = user_levels
 
-        # get all group-specific permissions
-        group_levels = {}
-        for rm in GroupObjectRoleMapping.objects.filter(object_id=self.id, object_ct=my_ct).all():
-            group_levels[rm.group.slug] = rm.role.codename
-        levels['groups'] = group_levels
+        if "geonode.contrib.groups" in settings.INSTALLED_APPS:
+            # get all group-specific permissions
+            group_levels = {}
+            for rm in GroupObjectRoleMapping.objects.filter(object_id=self.id, object_ct=my_ct).all():
+                group_levels[rm.group.slug] = rm.role.codename
+            levels['groups'] = group_levels
 
         return levels
 
@@ -353,15 +361,16 @@ class PermissionLevelMixin(object):
             user = User.objects.get(username=username)
             self.set_user_level(user, level)
 
-        #TODO: Should this run in a transaction?
-        excluded_groups = [g[0] for g in perm_spec.get('groups', list())]
+        if "geonode.contrib.groups" in settings.INSTALLED_APPS:
+            #TODO: Should this run in a transaction?
+            excluded_groups = [g[0] for g in perm_spec.get('groups', list())]
 
-        # Delete all group levels that do not exist in perm_spec.
-        self.get_group_levels().exclude(group__slug__in=excluded_groups).delete()
+            # Delete all group levels that do not exist in perm_spec.
+            self.get_group_levels().exclude(group__slug__in=excluded_groups).delete()
 
-        for group, level in perm_spec.get('groups', list()):
-            group = Group.objects.get(slug=group)
-            self.set_group_level(group, level)
+            for group, level in perm_spec.get('groups', list()):
+                group = Group.objects.get(slug=group)
+                self.set_group_level(group, level)
 
 # Logic to login a user automatically when it has successfully
 # activated an account:
